Applied patchs from Karvjorm (tickets #7125, #7131, #7126, #7129): Localizations for PNG PCX HVIF and HPGS translators.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@40262 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/translators/hpgs/ReadHelper.h
@@ -90,7 +90,7 @@ positionio_read(void *ptr, size_t size, size_t nmemb, my_hpgs_istream *stream)
 	unsigned char *iptr = (unsigned char *)ptr;
 	size_t i = 0;
 	for (; i < nmemb; i++) {
-		if (size != stream->buffer->Read(iptr, size))
+		if (size != (size_t)stream->buffer->Read(iptr, size))
 			break;
 		iptr += size;
 	}
